+++ R/scoringRules_wrappers.R
@@ -8,7 +8,7 @@
 #' to integer forecasts, the problem lies in the implementation: The Log Score
 #' needs a kernel density estimation, which is not well defined with
 #' integer-valued Monte Carlo Samples. The Log Score can be used for specific
-#' integer valued probabiliy distributions. See the scoringRules package for
+#' integer valued probability distributions. See the scoringRules package for
 #' more details.
 #' @param true_values A vector with the true observed values of size n
 #' @param predictions nxN matrix of predictive samples, n (number of rows) being